requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
soundfile>=0.12.1
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
